serial_w32: Flush input buffer on Windows too
From the beginning the actual flush was commented out and not adapted to
the Windows API since at the time there was no apparent need to flush.
Especially since the GPIO sequences were added flushing can be needed,
for instance after resetting into the bootloader from a talkative
application.

Thanks to Jan Belohoubek for reporting, fix, and testing.

https://sourceforge.net/p/stm32flash/tickets/132/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>