dev_table: Tighten option bytes range for device ID 0x436 / 0x437
These dual-bank devices have two ranges of option bytes, and we
previously listed the start of the first and the end of the second.
Since we use the range (only) to display the number of option bytes,
this can get very wrong for some dual-bank devices where the two ranges
are far apart. So for consistency ignore the second range, as long as
our table only support one range.

https://sourceforge.net/p/stm32flash/tickets/92/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>